refactor lp staking & add ga
